+++ src/textual/messages.py
@@ -29,7 +29,7 @@ class UpdateMessage(Message, verbosity=3):
 @rich.repr.auto
 class LayoutMessage(Message, verbosity=3):
     def can_replace(self, message: Message) -> bool:
-        return isinstance(message, LayoutMessage)
+        return isinstance(message, LayoutMessage) or isinstance(message, UpdateMessage)
 
 
 @rich.repr.auto