+++ src/textual/widget.py
@@ -40,6 +40,15 @@ if TYPE_CHECKING:
 log = getLogger("rich")
 
 
+class Spacing(NamedTuple):
+    """The spacing around a renderable."""
+
+    top: int = 0
+    right: int = 0
+    bottom: int = 0
+    left: int = 0
+
+
 class RenderCache(NamedTuple):
     size: Size
     lines: Lines
@@ -85,19 +94,19 @@ class Widget(MessagePump):
     layout_offset_y: Reactive[float] = Reactive(0.0, layout=True)
 
     style: Reactive[str | None] = Reactive(None)
-    padding: Reactive[PaddingDimensions | None] = Reactive(None, layout=True)
-    margin: Reactive[PaddingDimensions | None] = Reactive(None, layout=True)
+    padding: Reactive[Spacing | None] = Reactive(None, layout=True)
+    margin: Reactive[Spacing | None] = Reactive(None, layout=True)
     border: Reactive[str] = Reactive("none", layout=True)
     border_style: Reactive[str] = Reactive("")
     border_title: Reactive[TextType] = Reactive("")
 
     BOX_MAP = {"normal": box.SQUARE, "round": box.ROUNDED, "bold": box.HEAVY}
 
-    def validate_padding(self, padding: PaddingDimensions) -> tuple[int, int, int, int]:
-        return Padding.unpack(padding)
+    def validate_padding(self, padding: PaddingDimensions) -> Spacing:
+        return Spacing(*Padding.unpack(padding))
 
-    def validate_margin(self, padding: PaddingDimensions) -> tuple[int, int, int, int]:
-        return Padding.unpack(padding)
+    def validate_margin(self, padding: PaddingDimensions) -> Spacing:
+        return Spacing(*Padding.unpack(padding))
 
     def validate_layout_offset_x(self, value) -> int:
         return int(value)
@@ -169,6 +178,16 @@ class Widget(MessagePump):
     def layout_offset(self) -> tuple[int, int]:
         """Get the layout offset as a tuple."""
         return (round(self.layout_offset_x), round(self.layout_offset_y))
+
+    @property
+    def gutter(self) -> Spacing:
+        mt, mr, mb, bl = self.margin or (0, 0, 0, 0)
+        pt, pr, pb, pl = self.padding or (0, 0, 0, 0)
+        border = 1 if self.border else 0
+        gutter = Spacing(
+            mt + pt + border, mr + pr + border, mb + pb + border, bl + pl + border
+        )
+        return gutter
 
     def _update_size(self, size: Size) -> None:
         self._size = size